ticket/requirements.txt
Flask
python-dotenv
prometheus-client
psutil
gunicorn
confluent-kafka
Marshmallow
pymongo
retry
bs4
requests